// React/app/GerAdv_TS/ContatoCRMView/Interfaces/interface.ContatoCRMView.ts
'use client';

export interface IContatoCRMView {
// 202501251
    id: number;
 
	cguid: string,
	data: string,
	ip: string,
}

export interface IContatoCRMViewFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IContatoCRMViewIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}